+++ provider/aws_test.go
@@ -133,16 +133,15 @@ func (r *Route53APIStub) CreateHostedZone(input *route53.CreateHostedZoneInput) 
 
 func TestAWSRecords(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "list-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("list-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	records, err := provider.Records(testZone)
 	if err != nil {
 		t.Fatal(err)
 	}
-
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "list-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("list-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 }
 
@@ -150,7 +149,7 @@ func TestAWSCreateRecords(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{})
 
 	records := []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	if err := provider.CreateRecords(testZone, records); err != nil {
@@ -163,20 +162,20 @@ func TestAWSCreateRecords(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 }
 
 func TestAWSUpdateRecords(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	}
 	updatedRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "1.2.3.4"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "1.2.3.4", "A"),
 	}
 
 	if err := provider.UpdateRecords(testZone, updatedRecords, currentRecords); err != nil {
@@ -189,17 +188,17 @@ func TestAWSUpdateRecords(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "1.2.3.4"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "1.2.3.4", "A"),
 	})
 }
 
 func TestAWSDeleteRecords(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	if err := provider.DeleteRecords(testZone, currentRecords); err != nil {
@@ -216,23 +215,23 @@ func TestAWSDeleteRecords(t *testing.T) {
 
 func TestAWSApplyChanges(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	createRecords := []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 	updatedRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "1.2.3.4"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "1.2.3.4", ""),
 	}
 
 	deleteRecords := []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	changes := &plan.Changes{
@@ -252,8 +251,8 @@ func TestAWSApplyChanges(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "1.2.3.4"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "1.2.3.4", "A"),
 	})
 }
 
@@ -269,7 +268,7 @@ func TestAWSCreateRecordsDryRun(t *testing.T) {
 	provider := newAWSProvider(t, true, []*endpoint.Endpoint{})
 
 	records := []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	if err := provider.CreateRecords(testZone, records); err != nil {
@@ -286,14 +285,14 @@ func TestAWSCreateRecordsDryRun(t *testing.T) {
 
 func TestAWSUpdateRecordsDryRun(t *testing.T) {
 	provider := newAWSProvider(t, true, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 	updatedRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "1.2.3.4"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "1.2.3.4", ""),
 	}
 
 	if err := provider.UpdateRecords(testZone, updatedRecords, currentRecords); err != nil {
@@ -306,17 +305,17 @@ func TestAWSUpdateRecordsDryRun(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 }
 
 func TestAWSDeleteRecordsDryRun(t *testing.T) {
 	provider := newAWSProvider(t, true, []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	if err := provider.DeleteRecords(testZone, currentRecords); err != nil {
@@ -329,29 +328,29 @@ func TestAWSDeleteRecordsDryRun(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 }
 
 func TestAWSApplyChangesDryRun(t *testing.T) {
 	provider := newAWSProvider(t, true, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	createRecords := []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 	updatedRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "1.2.3.4"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "1.2.3.4", ""),
 	}
 
 	deleteRecords := []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", ""),
 	}
 
 	changes := &plan.Changes{
@@ -371,8 +370,8 @@ func TestAWSApplyChangesDryRun(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 }
 
@@ -380,7 +379,7 @@ func TestAWSCreateRecordsCNAME(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{})
 
 	records := []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", ""),
 	}
 
 	if err := provider.CreateRecords(testZone, records); err != nil {
@@ -393,20 +392,20 @@ func TestAWSCreateRecordsCNAME(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", "CNAME"),
 	})
 }
 
 func TestAWSUpdateRecordsCNAME(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", "CNAME"),
 	})
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", ""),
 	}
 	updatedRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "bar.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "bar.elb.amazonaws.com", ""),
 	}
 
 	if err := provider.UpdateRecords(testZone, updatedRecords, currentRecords); err != nil {
@@ -419,17 +418,17 @@ func TestAWSUpdateRecordsCNAME(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "bar.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "bar.elb.amazonaws.com", "CNAME"),
 	})
 }
 
 func TestAWSDeleteRecordsCNAME(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "baz.elb.amazonaws.com"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "baz.elb.amazonaws.com", "CNAME"),
 	})
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "baz.elb.amazonaws.com"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "baz.elb.amazonaws.com", ""),
 	}
 
 	if err := provider.DeleteRecords(testZone, currentRecords); err != nil {
@@ -446,23 +445,23 @@ func TestAWSDeleteRecordsCNAME(t *testing.T) {
 
 func TestAWSApplyChangesCNAME(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "qux.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", "CNAME"),
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "qux.elb.amazonaws.com", "CNAME"),
 	})
 
 	createRecords := []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", ""),
 	}
 
 	currentRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "bar.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "bar.elb.amazonaws.com", ""),
 	}
 	updatedRecords := []*endpoint.Endpoint{
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "baz.elb.amazonaws.com"},
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "baz.elb.amazonaws.com", ""),
 	}
 
 	deleteRecords := []*endpoint.Endpoint{
-		{DNSName: "delete-test.ext-dns-test.teapot.zalan.do", Target: "qux.elb.amazonaws.com"},
+		endpoint.NewEndpoint("delete-test.ext-dns-test.teapot.zalan.do", "qux.elb.amazonaws.com", ""),
 	}
 
 	changes := &plan.Changes{
@@ -482,14 +481,14 @@ func TestAWSApplyChangesCNAME(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "create-test.ext-dns-test.teapot.zalan.do", Target: "foo.elb.amazonaws.com"},
-		{DNSName: "update-test.ext-dns-test.teapot.zalan.do", Target: "baz.elb.amazonaws.com"},
+		endpoint.NewEndpoint("create-test.ext-dns-test.teapot.zalan.do", "foo.elb.amazonaws.com", "CNAME"),
+		endpoint.NewEndpoint("update-test.ext-dns-test.teapot.zalan.do", "baz.elb.amazonaws.com", "CNAME"),
 	})
 }
 
 func TestAWSSanitizeZone(t *testing.T) {
 	provider := newAWSProvider(t, false, []*endpoint.Endpoint{
-		{DNSName: "list-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("list-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	records, err := provider.Records(testZone)
@@ -498,7 +497,7 @@ func TestAWSSanitizeZone(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "list-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("list-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 
 	records, err = provider.Records("/hostedzone/" + testZone)
@@ -507,13 +506,13 @@ func TestAWSSanitizeZone(t *testing.T) {
 	}
 
 	validateEndpoints(t, records, []*endpoint.Endpoint{
-		{DNSName: "list-test.ext-dns-test.teapot.zalan.do", Target: "8.8.8.8"},
+		endpoint.NewEndpoint("list-test.ext-dns-test.teapot.zalan.do", "8.8.8.8", "A"),
 	})
 }
 
 func validateEndpoints(t *testing.T, endpoints []*endpoint.Endpoint, expected []*endpoint.Endpoint) {
 	if !testutils.SameEndpoints(endpoints, expected) {
-		t.Errorf("expected %v, got %v", expected, endpoints)
+		t.Fatalf("expected %v, got %v", expected, endpoints)
 	}
 }
 
@@ -552,7 +551,6 @@ func setupRecords(t *testing.T, provider *AWSProvider, endpoints []*endpoint.End
 	if err != nil {
 		t.Fatal(err)
 	}
-
 	validateEndpoints(t, records, endpoints)
 }