+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Features:
+  - Ownership via TXT records 
+    1. Create TXT records to mark the records managed by External DNS 
+    2. Supported for AWS Route53 and Google CloudDNS 
+    3. Configurable TXT record DNS name format 
+
 ## v0.2.0 - 2017-04-07
 
 Features: